# 8268229: Aarch64: Use Neon in intrinsics for String.equals