--- a/src/hotspot/cpu/aarch64/aarch64.ad
+++ b/src/hotspot/cpu/aarch64/aarch64.ad
@@ -16673,7 +16673,7 @@ instruct string_equalsL(iRegP_R1 str1, iRegP_R3 str2, iRegI_R4 cnt,
 
   format %{ "String Equals $str1,$str2,$cnt -> $result" %}
   ins_encode %{
-    // Count is in 8-bit bytes; non-Compact chars are 8 bits.
+    // Count is in 8-bit bytes.
     __ string_equals($str1$$Register, $str2$$Register,
                      $result$$Register, $cnt$$Register, 1);
   %}
@@ -16689,7 +16689,7 @@ instruct string_equalsU(iRegP_R1 str1, iRegP_R3 str2, iRegI_R4 cnt,
 
   format %{ "String Equals $str1,$str2,$cnt -> $result" %}
   ins_encode %{
-    // Count is in 8-bit bytes; non-Compact chars are 16 bits.
+    // Count is in 8-bit bytes.
     __ string_equals($str1$$Register, $str2$$Register,
                      $result$$Register, $cnt$$Register, 2);
   %}