--- a/src/hotspot/cpu/aarch64/globals_aarch64.hpp
+++ b/src/hotspot/cpu/aarch64/globals_aarch64.hpp
@@ -93,6 +93,8 @@ define_pd_global(intx, InlineSmallCode,          1000);
           "Use SIMD instructions in generated array equals code")       \
   product(bool, UseSimpleArrayEquals, false,                            \
           "Use simpliest and shortest implementation for array equals") \
+  product(bool, UseSimpleStringEquals, true,                            \
+          "Use simpliest and shortest implementation for string equals")\
   product(bool, UseSIMDForBigIntegerShiftIntrinsics, true,              \
           "Use SIMD instructions for left/right shift of BigInteger")   \
   product(bool, AvoidUnalignedAccesses, false,                          \