--- a/src/hotspot/cpu/aarch64/macroAssembler_aarch64.cpp
+++ b/src/hotspot/cpu/aarch64/macroAssembler_aarch64.cpp
@@ -4763,7 +4763,7 @@ address MacroAssembler::arrays_equals(Register a1, Register a2, Register tmp3,
 void MacroAssembler::string_equals(Register a1, Register a2,
                                    Register result, Register cnt1, int elem_size)
 {
-  Label SAME, DONE, SHORT, NEXT_WORD;
+  Label SAME, DONE, SHORT, NEXT_WORD, MEDIUM, NEXT_TWO_WORDS;
   Register tmp1 = rscratch1;
   Register tmp2 = rscratch2;
   Register cnt2 = tmp2;  // cnt2 only used in array length compare
@@ -4782,8 +4782,33 @@ void MacroAssembler::string_equals(Register a1, Register a2,
 
   mov(result, false);
 
+  // Check for short strings, i.e. smaller than wordSize * 2.
+  subs(cnt1, cnt1, wordSize * 2);
+  br(Assembler::LT, MEDIUM);
+  // Main 16 byte comparison loop.
+  bind(NEXT_TWO_WORDS); {
+    ldr(v0, Q, Address(post(a1, wordSize * 2)));
+    ldr(v1, Q, Address(post(a2, wordSize * 2)));
+    subs(cnt1, cnt1, wordSize * 2);
+    eor(v0, T16B, v0, v1);
+    mov(tmp1, v0, T2D, 0);
+    cbnz(tmp1, DONE);
+    mov(tmp2, v0, T2D, 1);
+    cbnz(tmp2, DONE);
+  } br(GT, NEXT_TWO_WORDS);
+  // Same as MEDIUM
+  ldr(v0, Q, Address(a1, cnt1));
+  ldr(v1, Q, Address(a2, cnt1));
+  eor(v0, T16B, v0, v1);
+  mov(tmp1, v0, T2D, 0);
+  cbnz(tmp1, DONE);
+  mov(tmp2, v0, T2D, 1);
+  cbnz(tmp2, DONE);
+  b(SAME);
+
   // Check for short strings, i.e. smaller than wordSize.
-  subs(cnt1, cnt1, wordSize);
+  bind(MEDIUM);
+  adds(cnt1, cnt1, wordSize);
   br(Assembler::LT, SHORT);
   // Main 8 byte comparison loop.
   bind(NEXT_WORD); {